2011-03-28 Roman - SYS: Using "pkg-config" for global include flag detect for libpcre is added. - SYS: To user's API object TCntrNode is added function nodePrev() for access parent node access. - SYS: To user's API object XMLNodeObj is added function parent() for access parent node access. - !!!:TFunction, DAQ.JavaLikeCalc: User's API object TFuncArgObj is added for get a function arguments. - DAQ.JavaLikeCalc: The special attribute "this" is added for user access the object of the controller allow. - DAQ.LogicLev: The special attribute "this" is added for user access the object of the controller's parameter allow. - FIX:DAQ.OPC_UA: Archive's value period set is fixed on an attribute to archive enabling. - !!:DAQ.SNMP: All value types support is added. Write values support is added. Errors processing is improved.     Addon SNMP connect parameters is added.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1315 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/Vision/vis_shapes.h
@@ -445,11 +445,11 @@ class ShapeProtocol : public WdgShape
 		//Methods
 		ShpDt( ) : active(true), timeCurent(false), time(0), tSize(60), arhBeg(0), arhEnd(0)	{ }
 		//Attributes
-		unsigned active		:1;
-		unsigned timeCurent	:1;
-		unsigned trcPer		:10;
-		unsigned lev		:4;
-		unsigned viewOrd	:4;
+		short active	:1;
+		short timeCurent:1;
+		short trcPer	:10;
+		short lev	:4;
+		short viewOrd	:4;
 		QTimer 		*trcTimer;
 		QTableWidget	*addrWdg;
 		unsigned int	time, tSize;
